requirements.txt
  matplotlib
  numpy
  pyyaml
  scikit-image
  scipy